sphinx_ext: move sphinx extensions to their own file
