+++ setup.cfg
@@ -55,3 +55,6 @@ ignore_missing_imports = True
 
 [mypy-whoosh.*]
 ignore_missing_imports = True
+
+[mypy-sphinx_click.*]
+ignore_missing_imports = True
